+++ components/molecules/index.ts
@@ -1,9 +1,13 @@
 import ClubItem from './ClubItem';
 import ClubList from './ClubList';
 import Gnb from './Gnb';
+import SelectMaxPeople from './SelectMaxPeople';
+import UploadPhotos from './UploadPhotos';
 
 export default {
   ClubList,
   ClubItem,
-  Gnb
+  Gnb,
+  SelectMaxPeople,
+  UploadPhotos
 };
